+++ src/dataframerow/dataframerow.jl
@@ -635,5 +635,9 @@ function Base.push!(df::DataFrame, dfr::DataFrameRow; cols::Symbol=:setequal,
 end
 
 metadata(dfr::DataFrameRow) = metadata(parent(dfr))
-metadata!(dfr::DataFrameRow, p; mode::Symbol) =
-    metadata!(parent(dfr), p, mode=mode)
+hasmetadata(dfr::DataFrameRow) = hasmetadata(parent(dfr))
+
+metadata(dfr::DataFrameRow, col::ColumnIndex) =
+    metadata(parent(dfr), _names(dfr)[index(dfr)[col]])
+hasmetadata(dfr::DataFrameRow, col::ColumnIndex) =
+    hasmetadata(parent(dfr), _names(dfr)[index(dfr)[col]])